+++ src/api/login.js
@@ -2,8 +2,8 @@ import axios from 'axios';
 
 export const login = async (data) => {
     try {
+        const response = await axios.post('https://www.silvercareer.shop/api/login', data);
         //const response = await axios.post('https://www.silvercareer.shop/api/login', data);
-        const response = await axios.post('http://localhost:8888/api/login', data);
         return response;
     } catch (error) {
         console.error(error);
